Tercer commit

+++ prestamo/src/main/java/com/tps/prestamo/models/Prestamo.java
@@ -10,21 +10,21 @@ import javax.persistence.GenerationType;
 public class Prestamo {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Integer id;
-    private Integer id_solicitante;
-    private Integer id_prestamista;
-    private Integer id_bookitem;
+    private Long id;
+    private Long id_solicitante;
+    private Long id_prestamista;
+    private Long id_bookitem;
     private Date fecha_inicio;
     private Date fecha_termino;
     
 
-    public void setIdSolicitante( Integer id_solicitante){
+    public void setIdSolicitante( Long id_solicitante){
         this.id_solicitante = id_solicitante;   
     }
-    public void setIdPrestamista (Integer id_prestamista){
+    public void setIdPrestamista (Long id_prestamista){
         this.id_prestamista = id_prestamista;
     }
-    public void setIdBookItem( Integer id_bookitem){
+    public void setIdBookItem( Long id_bookitem){
         this.id_bookitem = id_bookitem;
     }
     public void setFechaInicio (Date fecha_inicio){
@@ -34,13 +34,13 @@ public class Prestamo {
         this.fecha_termino = fecha_termino;
     }
 
-    public Integer getIdSolicitante(){
+    public Long getIdSolicitante(){
         return this.id_solicitante;
     }
-    public Integer getIdPrestamista(){
+    public Long getIdPrestamista(){
         return this.id_prestamista;
     }
-    public Integer getIdBookItem(){
+    public Long getIdBookItem(){
         return this.id_bookitem;
     }
     public Date getFechaInicio(){
