+++ prestamo/src/main/java/com/tps/prestamo/repository/PrestamoRepository.java
@@ -2,6 +2,6 @@ package com.tps.prestamo.repository;
 import com.tps.prestamo.models.Prestamo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PrestamoRepository extends JpaRepository<Prestamo,Integer> {
+public interface PrestamoRepository extends JpaRepository<Prestamo,Long> {
 
 }
